Refactored the patterns part of the index, fixed patterns loading bugs.
