+++ src/index/CMakeLists.txt
@@ -13,7 +13,9 @@ add_library(index
         DocsDB.cpp DocsDB.h
         FTSAlgorithms.cpp FTSAlgorithms.h
         PrefixHeuristic.cpp PrefixHeuristic.h
-        PatternContainer.cpp PatternContainer.h)
+        PatternContainer.cpp PatternContainer.h
+        PatternIndex.cpp PatternIndex.h
+        VocabularyData.h)
 
 target_link_libraries(index parser ${STXXL_LIBRARIES} ${ICU_LIBRARIES} absl::flat_hash_map absl::flat_hash_set)
